pyproject.toml
[tool.poetry]
name = "repl_python3_Jerry-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
Flask = "^2.0.2"
replit = "^3.2.4"
firebase-admin = "^5.2.0"
pymongo = "^4.0.1"
dnspython = "^2.1.0"
captcha = "^0.3"
Pyrebase4 = "^4.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
